# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Secretary API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/professors": {
            "get": {
                "tags": [
                    "Professors"
                ],
                "summary": "Lister tous les professeurs",
                "operationId": "3ad078f5b9cc0657748a4099913b5c58",
                "responses": {
                    "200": {
                        "description": "Liste des professeurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "Get Professors"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "firstname": {
                                                        "type": "string",
                                                        "example": "Jean"
                                                    },
                                                    "lastname": {
                                                        "type": "string",
                                                        "example": "Dupont"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "jean@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Professors"
                ],
                "summary": "Créer un professeur",
                "operationId": "0f47f8689f6a1643e05b80359c4dcf05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jean@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Professeur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "Professor Storage"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Professeur creé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "professor": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "firstname": {
                                                            "type": "string",
                                                            "example": "Jean"
                                                        },
                                                        "lastname": {
                                                            "type": "string",
                                                            "example": "Dupont"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "jean@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/professors/{professor}": {
            "put": {
                "tags": [
                    "Professors"
                ],
                "summary": "Modifier un professeur",
                "operationId": "e07afc1aedf6ebddfbd65c4c9c065a2a",
                "parameters": [
                    {
                        "name": "professor",
                        "in": "path",
                        "description": "ID du professeur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jean@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Professeur modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "Professor Update"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Informations du professeur modifiées avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "professor": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "firstname": {
                                                            "type": "string",
                                                            "example": "Jean"
                                                        },
                                                        "lastname": {
                                                            "type": "string",
                                                            "example": "Dupont"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "jean@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Professors"
                ],
                "summary": "Supprimer un professeur",
                "operationId": "7382d7fa2fd124fac7f5da65ba1a2a7d",
                "parameters": [
                    {
                        "name": "professor",
                        "in": "path",
                        "description": "ID du professeur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Professeur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "Professor Delete"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Professeur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}